HREPTWO-4536: split gallery and folder workflow plugins
The gallery workflow plugin now only takes care of uploading images & assets.
It has it's own category, for the image and asset sections in the browser.
Having different kinds of asset folders is a matter of configuration; creating
a separate template under /hippo:configuration/hippo:queries/hippo:templates
for the foldertype should be sufficient.