CMS7-5080: use array access for array of elements
The rows in the foot of the table were accessed as an element instead of an
array.  This borked the height calculation algorithm.  Note that this does not
work in IE, it didn't work in the past either so it's not a regression.
(scrolling tables in IE are a notorious problem)