CMS7-5032 SSO HST Preview from Channel Manager
This commit adds the CMS side of code  to make use CMS as SSO authentication provider. There's a new mount "/auth" available in the CMS now. 

The (web)applications which want to use the CMS as the authentication provider  can redirect  to http://CMS_LOCATION/auth?key=SOMEKEY&destinationUrl=DESTINATION_URL, where the KEY is unique key that the application generated.
 
 The destinationUrl should be the URL where the CMS should post the credentials to. If the user is logged into CMS, then CMS will redirect to the destination URL and will provide the encrypted credentials in the URL parameter cred. CMS uses cms-sso module to encrypt the credentials. 
 
 Please note that currently, the *same instance* of org.onehippo.sso.CredentialCipher must be used to decryp the credentials.