CMS7-3953: switch to hippo-cms-api as the primary dependency to build plugins to
This checkin contains a compiling baseline, the changes still need to be
validated to a full project.  These do not include updated cms poms beyond
the minimal changes necessary to get a compiling build.

The hippo-cms-engine now implements the functionality that is offered by the
API.  Anything that's in the engine should be considered out-of-bounds for
plugins.

The yui module has been merged into the api module, as there were some
components already in the api that used yui.

All api tests have been moved into the test module.  It is now a mix of
api-level functionality and engine tests.

The compressor plugin in the jquery module has been replaced by the hippo
(yui) compressor.  Maven was not able to resolve the alchim one currently.

Login modules have been moved into the engine.  This hides credentials from
plugin developers, they can still retrieve the user name from the jcr session.