CMS7-5711: The delivered patch does something fundamentally different, which might also be what happened in the past.  Instead of using the root session in the workflow manager to execute a linked workflow call, the workflow manager of the root session is accessed and the workflow is executed in the root session of that workflowmanager.  So in that case there are three sessions working, the original from the user, the root session created by the workflowmanager of that user session, and a root-root session.
The problem however is that in the TemplateEditorWorkflowImpl two other
workflows are invoked, first a call to create the namespace and then
one to create a folder structure.  However these However the first
is executed using workflow chaining, while the second using indirect
workflow calls, which means that the second is called before the first,
leading to an exception.