CMS7-5770: resolve document for a version
When a document is published, it is also versioned.  Since the workflow manager
now returns the last returned object in a workflow chain, the version is now
returned and put in the event log.

The event model now tries to resolve the document that corresponds to the
version when encountering a version in the event log.