CMS7-5817: added search to the Users and the Groups panel
Both the list of users and the list of groups can be searched. All terms are ANDed, unless separated by 'or'. The * and ? can be used as wildcards for multiple and single characters. A - before a term will exclude instead of include it. Obvious queries that will return all users/groups ("*", "* or *", etc.) are filtered out.

Also refactored the UsersDataProvider and GroupsDataProvider to use a common abstract class that implements the search functionality.