CMS7-5823: use HippoEvent implementation from hippo-cms7-commons
The hippo-cms7-commons provides the hippo event and will offer additional functionality based on it in the future.
The auditlogger has been removed from the API as it is not intended to be used by plugins.  The two places that need it within the cms (login plugin & admin perspective) now look the slf4j logger up themselves.