Backported:
Revision #33329 Committed by wdanes 3/7/12 5:39 PM CMS7-5907: Moved the code that creates the group links to it's own panel, so Collections of Wicket Components aren't passed around CMS7-5927: Moved the code that creates the domain links to it's own panel, so Collections of Wicket Components aren't passed around  Also removed the AjaxLinkLabelListPanel, because it's no longer used due to the two changes above.  CMS7-5927: The user detail screen now shows which groups have which roles on which domains instead of just a list of roles and domains that are somehow coupled with the group.  Readded the "Type" colum for the user list screen.