CMS7-5875: refactor selected extension point logic
The selected extension point no longer needs to be explicitly updated by plugin
editor plugins.  Instead, the ILayoutAware interface has been enriched to allow
the document/compound type editor to select an extension point by itself.

The pads in a layout descriptor are now ordered; the first pad is one selected
by default.

The inherited field section now also shows fields that only have an editor for
a subtype.  I.e. when a type does not have an editor for a field, but the
editor for a sub-type does, the field will still be suggested in the
sub-sub-type builder.