CMS7-6165: Use JCR observations to invalidate user & group data provider caches
The list panels are now observers, registering when they are added to the breadcrumb bar and unregistering when removed.

The user list panel is now no longer shown by default when a user is deleted. When the user details were shown, the previous panel will be activated. When deleting from the list of users, the list will remain active.