CMS7-6440 REGRESSION: When opening a revision, the wrong menu is displayed
AbstractCmsEditor registers two IModelReferenceServices when starting the Editor cluster, one with the current node as model and the other with the versioned node for comparison. The editor in this case is opened via the HistoryDialog, when the cluster gets started the WorkflowsPlugin which renders and builds the workflow action menu loads in it's super class AbstractRenderService the model via the model service in the constructor, which when it isn't configured as the reverenced model.compareTo is the default reference wicket.model which refers to the baseService registered previously by the AbstractCmsEditor. Using the model.compareTo for the WorkflowsPlugin does invoke the right service and fixes the regression.