CMS7-6699: fix stacktraces and non-flash upload breakage
The service contexts were stopped prematurely; some repeaters (listview, refreshingview) rebuild the component hierarchy and invoke onRemove callbacks.  The same component may be added to the hierarchy just after having been removed.  So the service context is now reattached when this happens.

The tabbed panel no longer fires as many remove calls.  Those traverse the entire hierarchy, which is quite expensive in the case of perspectives.

The onSubmit method in the abstract dialog no longer tries the find the submitting button.  Since it may have already been removed from the hierarchy in a handleSubmit invocation, this is not safe.  (the form will throw an exception when it finds the button but it is no longer visible)