CMS7-6360 - IE8: when uploading a document to assets an JS error is thrown. (Flash)
In some cases the close routine of the dialog is called twice, this will cause the second call to fail and throw a javascript error. For example, clicking OK in the unpublish dialog will first call close as part of the onclick handling of the OK/Cancel button, and afterwards close is called again by the DialogServiceFactory.releaseService routine.
For assets however, if we select folder A, then 'right-click' on folder B and upload an asset, the DialogServiceFactory.releaseService will execute first, and later on the onclick handler of the OK button is called.
Both paths are now safeguarded against executing twice for the same dialog.

Also, nested dialogs (pending dialogs in the code) had problems with attaching/detaching the keyboard handlers: the handlers of the new Dialog would be attached before the old ones were removed.