CMS7-7092 Renaming a folder doesn't create a correctly localised hippo:translation node
When a folder is created no locale is provided for the hippo:translslation node which results in the same node message is overwritten by users logging in with different languages renaming the folder. The user session locale is now used when a folder is created, if a folder is renamed the message for the user session locale is used to determine weather renaming is necessary.