CMS7-7265: initial integration with CKEditor
A separate editor instance is loaded for each HTML field, and destroyed when the HTML element is removed. 
The associated textarea of each editor is updated whenever the onChange plugin of the editor generates
an 'onchange' event. The contents of the textarea is then submitted when the document is saved. Setting
the 'language' configuration parameter ensures that the editor UI is localized in the current CMS locale. 

Note that 'compare' mode is not implemented yet.