CMS7-7277: added CKEditor plugin 'hippoimagepicker' that opens the Hippo image picker dialog.
The ckeditor module has been changed to a pom module with two sub-modules: ckeditor-resources with the ckeditor sources as built by ckeditor.com/builder, and ckeditor-plugins with Hippo-specific plugins for CKEditor.  

The Xinha image picker dialog code is largely copied from the Xinha source tree and modified in a few places to integrate with CKEditor instead of Xinha. Some Xinha-specific classes have been made generic and moved to the richtext source tree.

The general flow is similar to Xinha: the hippoimagepicker plugin executes a Wicket Ajax call to open the image picker dialog. Meta-data stored in the img tag is send along to open the right image in the dialog. When the dialog is closed, the 'insertImage' command of the hippoimagepicker plugin is executed to insert (or replace) the picked image.