CMS7-7307: added styles to the 'styles' toolbar, made contents CSS pluggable again.
The contents.css file in CKEditor can be tweaked to style the editor contents. That same file is used to style the elements in the 'styles' combo. 
Since we use the divarea plugin, we lost the ability to style the editing area, but we needed that again to ensure that the styles of the elements 
in the 'styles' combo look the same in the editing area. For this reason we now include a hippocontents.css by default. All styling in there
should be scoped to .cke_wysiwyg_div.cke_editable for the editing area, or .cke_panel_container for menu combos. 

CKEditor style sets are not localized, so we create a different style set per language (hippo_en, hippo_nl, etc.) and use the one for the current locale.