CMS7-7315: removed the AutoGrow plugin
It does not play nice with the 'divarea' plugin, see http://dev.ckeditor.com/ticket/9351. The basic functionality works (the editor height grows with the content), but the configuration properties 'autoGrow_minimumHeight' and 'autoGrow_maximumHeight' are ineffective. As a result, the editor height grows indefinitely, even to the point where the toolbar is no longer visible.