CMS7-7377: added a way to add multiple elements to an existing array
Expanding an existing array is very common for the CKEditor configuration option 'toolbarGroups', 
where multiple toolbar group objects have to be added to an existing array of toolbar group objects.
 
Concatenating arrays is a choice, and removes the ability to append an array as an element of an existing array.