CMS7-7541: fix bookmarkability of urls on windows that have already been initialized
Windows that had a login screen initialized wouldn't pick up the parameters.  There was also some odd behavior due to the page object being created on log-in having the same id as the one used for the login screen.

The path is now also processed in the login request cycle; i.e. the opening of the correct document is not dependent on a popstate event. 

Bonus: back & forward navigation works when switching between tabs