CMS7-7575: cleanup dialog window resources when dialog is closed by a non-Ajax call
Some dialogs use a non-Ajax submit (e.g. XML import and CND import in the console), 
but in that case the dialog window did not cleanup its resources. Subsequent calls 
to the dialog window would then re-render the last content of the previously closed 
dialog, which did not trigger any UI change anymore.

Also cleaned up some logic in AbstractDialog. The superclass already calls 
setOutputMarkupId, and the isRemoved field already reflects whether the dialog is 
still on the page or not.