CMS7-7652 Fix hidden page navigation label.
Page-navigation buttons were at the right side of the table previous to 7.9.
Currently they are on the left side. As a consequence, the page navigation
label is no longer visible. Niels helped me out in fixing this by changing
the css a little, so that the label is now just  above the nav buttons.