CMS7-7707: let workflow descriptor model be able to obtain workflow
This allows the model to use the handle to look up the workflow when the subject is a frozen node.
Caching the workflow in the descriptor model also consolidates workflow lookup for all workflow plugins in a plugin cluster.
A nice benefit is that some boilerplate code could be removed in a number of other places.

The model constructor with an explicit descriptor has been deprecated; since within the CMS there is little use of the descriptor
without also needing a model, a constructor with both object and reconstruction parameters is only confusing.  (it can also lead
to the situation where the model is unable to re-resolve the descriptor)

The property-based frontend renderer is no longer supported.  It only introduced a separate class of "plugin" that was not used/tested
 in the certified stack.  Also the menu-item-for-workflow-method magic has been removed.  All UI elements should come from a plugin.