CMS7-7699: do not remove element around selection when picked image does not have a UUID
A picked image used have no UUID when the 'remove' button in the image picker dialog was clicked, 
but that 'remove' bbutton has been removed some time ago. However, the UUID is also empty when 
the 'close' button of the image picker dialog is clicked. If there was no image selected, the enclosing 
element (e.g. the enclosing paragraph) would be removed instead. That should not happen.