CMS7-7950: allow upload of .svg images
The dimensions of SVG images are not validated in the image upload plugin since processing them does not require a lot of memory. The svg files are processed separately by the scaling gallery processor, which reads the width and height from XML and adjusts it per variant. The image display plugin now also renders the 'width' and 'height' per image.