CMS7-7963 Do not enforce taggable node type
In general, the node being edited does not have to be of type mixin. Most 
likely a descendant node of the primaryNodeType will have this mixin.
For example in the archetype, the descendant node hipposysedit:prototype 
of the node gettingstarted:newsdocument has as mixin hippostd:taggable.
However, if you create a newsdocument, the document will NOT have the
mixin hippostd:taggable.