CMS7-8030 Remove check on nodetype nt:handle
, but they do need the
path parameter to be set. When opening a document type in the CMS the
path to the document type should be visible in the URL. Since document
types do not have nt:handle as node type the code should not check for
presence of such a type. That's why the if statement has been removed.