CMS7-8018 Make new document types selectable when creating new document
It is sufficient to remove the hippo:skipindex mixin from the 
hipposysedit:prototype as stored under hippo:configuation and make 
sure that the node gets reloaded. For completeness all other mixins 
will be deleted too. It would be nice if the node gets the mixins 
from hipposysedit:supertype upon creation, but I don't see a simple
way to do this.