CHANNELMGR-152: ensure that the resize handlers of modals do not overlap with a scrollbar in the body
The border is now changed back to a background shining through margins on all sides (i.e. the default ExtJS way).
All resize handlers are still 6px wide (the default), except the right-hand side one to avoid overlap with the scrollbar.

Also split the styling of the ghost and the real window to avoid generating unnecessary classes for the ghost window.