CMS7-8771: HippoWidget registers resize listener too soon/before the correct layout unit has been created. This causes two types of issues: - During a full page reload the WidgetBehavior will fire WidgetManager.render() while the layout is still rendering. This means the widget will not have a layout unit to provide size info and will fallback to guessing a correct size, mostly not what is intended. - During an ajax request the WidgetManager.render() method is also executed while the correct layout unit still needs to be rendered. In some cases this will lead to weird issues where the second ancestor layout unit will provide size info instead of the first ancestor. When the widget is then rendered again, it will register a second resize listener on the correct layout unit but due to the order which the event listener are called the incorrect size info is still the one that will be used.
To fix this a simple delay of execution is added for the registerResizeListener and registerRenderListener methods. All register calls are now stored in a FunctionQueue that is processed after the render phase has completed.
This lead to a simplification of the HippoWidget render calls. Instead of relying on a serverside call to "YAHOO.hippo.WidgetManager.render()", a HippoWidget register a render event listener on the layout manager.
To work around the fact that a dialog does not live inside a layout, it's render callback will be executed once if executeNow is set to true.