CMS7-9260: let a perspective generate a 'perspective-activated' event when it's activated
A perspective only generates events when an event ID is passed to its constructor. The event 
parameter 'perspectiveId' is set to this event ID.

The PerspectiveActivatedHeaderItem ensures that a 'perspective-activated' event is also
generated when the perspective is directly active during the first page load.

The scripts added by the root plugin have been refactored in two separate header items: one
for the root plugin itself, and one for scripts related to usage statistics. The usage statistics
initialization now generates the 'login' event. The generation of the 'logout' event has been
moved to the UserMenu.js script so it's closer to the actual markup.