CMS-9517: Make the link/unlink button look like a button
I also removed the left-padding on the disabled button as it has no background image. There are nice link/unlink icons in the Hippo Theme, but Ext wants a background image and I did not want to fork the two icons. We might want to move the button into the right side of the top-panel, clearing up pixels to be used for the second panel which is a bit cramped at the moment, especially the table/tree view.