CMS-9114: add active logout to the console
The Ext Ajax monitoring has been split off from the user-activity.js script and is only included in the CMS. The user activity API (and required monitoring JS) has been moved to the RootPlugins of the CMS and console, so the ActiveLogoutPlugin can be reused in both.
  
The logout logic of the console has been moved to a separate service. It's config node should be placed before the root plugin and the logout plugin, otherwise they cannot find the logout service.