CMS-9780 render field hints as title attributes
It is not possible to create CSS-only tooltips that are always
positioned correctly and adapt their size to the amount of text
shown. Instead the tooltip text is now put in a 'title' attribute,
which the browser will show in a nicely positioned tooltip. The
styling of these tooltips (white text on a black background) is
also more in line with the custom tooltips in other perspectives.