CMS-10100 Add <wicket:child/> to CaptchaForm
Add a <wicket:child/> element so subclasses of the CaptchaForm can extend the HTML instead of having to fork. The extension point is inserted *before* the captcha field so that when the Captcha is shown (after a couple of failed login attempts), it does not break the form layout.

(cherry picked from commit 7e7f005ede4b7824146dba7e4428408a368e4268)