CMS-10554  Use correct host in case request is delegated by a proxy like a CDN
Instead of having the 'Host' header as fallback for the 'X-Forwarded-Host' have the
'X-Forwarded-Host' header as fallback of 'Host' header. 'Host' header is since HTTP/1.1
mandatory and we have documented everywhere that this 'Host' header must be preserved
when running with proxies like httpd / haproxy / nginx, etc.

Next to that, small trivial code improvement that if getRenderingHost is not null, we
directly return: The getRenderingHost never returns a comma separated hosts string
but always a single host and is only used internally between channel mngr and HST
communication