CMS-10541 Enable 'copy formatting' button by default
The 'copy formatting' button is new in CKEditor 4.6. It seems very
useful, so enable it by default.

The inclusion of the toolbar group 'cleanup' in the
CKEditorNodePlugin also enables the 'remove format' button for
rich text fields. This button was already enabled by default in the
CKEditorPropertyPlugin (i.e. for formatted text fields), so this
aligns the toolbar of both plugins a little more.