CMS-10619 Fix NPE when retrieving class name
Class#getCanonicalName returns null if "it is a local or anonymous
class or an array whose componenttype does not have a canonical name)."
In that case Class#getName should be used as fallback.

Apart from that I did a bit of performance testing and it seems that
the check if the userIsLoggedIn is almost always faster than the
isWhitelisted check (probably because the getCanonicalName can be a
bit expensive).

(cherry picked from commit 0bc8127c6ed68dcfe55ac701bf2f131c8d2d9f43)