CMS-10619 Fix bootstrap configuration for CMS settings node
The settings node was backported to 7.9, where it did not exist
before. In there it only contains the property
'whitelisted.classes.for.package.resources'. To be sure that the
'send.usage.statistics.to.hippo' property is added during an
upgrade, both properties have to be added via delta XML, but only
when they did not exist already (which could mean their value was
changed by a customer).